create AccountDetailDictionary and use in TransactionTable
