+++ pages/index.tsx
@@ -3,23 +3,11 @@ import { useGetTransactionsTransactionsGet } from "../api/service/transactions";
 import { Tabs } from "../components/Tabs";
 import { Layout } from "../components/Layout";
 import { format, subDays } from "date-fns";
-import { useGetAccountsAccountsGet } from "../api/service/accounts";
-
-function MakeAccountDictionary() {
-  const response = useGetAccountsAccountsGet();
-  const mainAccounts = response.data?.main_accounts;
-  let accountDictionary = {};
-  if (mainAccounts) {
-    for (const mainAccount of mainAccounts) {
-      for (const account of mainAccount.accounts) {
-        accountDictionary[account.accountUid] = account.name;
-      }
-    }
-  }
-  return accountDictionary;
-}
+import MakeAccountDetailDictionary from "../utils/MakeAccountDetailDictionary";
 
 export default function Home() {
+  const accountDictionary = MakeAccountDetailDictionary();
+
   const end_date = new Date();
   const start_date = subDays(end_date, 28);
   const start_date_string = format(start_date, "yyyy-MM-dd'T'HH:mm:ss'Z'");
@@ -34,8 +22,6 @@ export default function Home() {
       query: { refetchInterval: 5 * 60 * 1000 },
     }
   );
-
-  const accountDictionary = MakeAccountDictionary();
 
   return (
     <>
